WIP: Avoid StmtNode assert
