+++ parser/parser_test.go
@@ -1773,25 +1773,6 @@ func (s *testParserSuite) TestInsertStatementMemoryAllocation(c *C) {
 	c.Assert(int(newStats.TotalAlloc-oldStats.TotalAlloc), Less, 1024*500)
 }
 
-func BenchmarkParse(b *testing.B) {
-	var table = []string{
-		"insert into t values (1), (2), (3)",
-		"insert into t values (4), (5), (6), (7)",
-		"select c from t where c > 2",
-	}
-	parser := New()
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		for _, v := range table {
-			_, err := parser.Parse(v, "", "")
-			if err != nil {
-				b.Failed()
-			}
-		}
-	}
-	b.ReportAllocs()
-}
-
 func (s *testParserSuite) TestExplain(c *C) {
 	defer testleak.AfterTest(c)()
 	table := []testCase{
